slave/requirements.txt
psycopg2-binary
flask
Flask-SQLAlchemy
gunicorn
requests
pytz
python-decouple
celery
